+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/android/MainActivity.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
 import io.sentry.Attachment;
 import io.sentry.ISpan;
+import io.sentry.MeasurementUnit;
 import io.sentry.Sentry;
 import io.sentry.SpanStatus;
 import io.sentry.UserFeedback;
@@ -27,6 +28,7 @@ import timber.log.Timber;
 public class MainActivity extends AppCompatActivity {
 
   private int crashCount = 0;
+  private int screenLoadCount = 0;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -200,8 +202,10 @@ public class MainActivity extends AppCompatActivity {
   @Override
   protected void onResume() {
     super.onResume();
+    screenLoadCount++;
     final ISpan span = Sentry.getSpan();
     if (span != null) {
+      span.setMeasurement("screen_load_count", screenLoadCount, new MeasurementUnit.Custom("test"));
       span.finish(SpanStatus.OK);
     }
   }