+++ sentry/src/main/java/io/sentry/SentryTracer.java
@@ -1,5 +1,6 @@
 package io.sentry;
 
+import io.sentry.protocol.MeasurementValue;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
 import io.sentry.protocol.TransactionNameSource;
@@ -13,6 +14,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Timer;
 import java.util.TimerTask;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
@@ -72,8 +74,9 @@ public final class SentryTracer implements ITransaction {
       new SpanByTimestampComparator();
   private final @NotNull AtomicBoolean isFinishTimerRunning = new AtomicBoolean(false);
 
-  private @Nullable TraceContext traceContext;
+  private final @NotNull Baggage baggage;
   private @NotNull TransactionNameSource transactionNameSource;
+  private final @NotNull Map<String, MeasurementValue> measurements;
 
   public SentryTracer(final @NotNull TransactionContext context, final @NotNull IHub hub) {
     this(context, hub, null);
@@ -104,6 +107,7 @@ public final class SentryTracer implements ITransaction {
       final @Nullable TransactionFinishedCallback transactionFinishedCallback) {
     Objects.requireNonNull(context, "context is required");
     Objects.requireNonNull(hub, "hub is required");
+    this.measurements = new ConcurrentHashMap<>();
     this.root = new Span(context, this, hub, startTimestamp);
     this.name = context.getName();
     this.hub = hub;
@@ -112,6 +116,12 @@ public final class SentryTracer implements ITransaction {
     this.trimEnd = trimEnd;
     this.transactionFinishedCallback = transactionFinishedCallback;
     this.transactionNameSource = context.getTransactionNameSource();
+
+    if (context.getBaggage() != null) {
+      this.baggage = context.getBaggage();
+    } else {
+      this.baggage = new Baggage(hub.getOptions().getLogger());
+    }
 
     if (idleTimeout != null) {
       timer = new Timer(true);
@@ -354,37 +364,44 @@ public final class SentryTracer implements ITransaction {
         // if it's an idle transaction which has no children, we drop it to save user's quota
         return;
       }
-      hub.captureTransaction(transaction, this.traceContext, null, profilingTraceData);
+
+      transaction.getMeasurements().putAll(measurements);
+
+      hub.captureTransaction(transaction, traceContext(), null, profilingTraceData);
     }
   }
 
   @Override
   public @Nullable TraceContext traceContext() {
     if (hub.getOptions().isTraceSampling()) {
-      synchronized (this) {
-        if (traceContext == null) {
-          final AtomicReference<User> userAtomicReference = new AtomicReference<>();
-          hub.configureScope(
-              scope -> {
-                userAtomicReference.set(scope.getUser());
-              });
-          this.traceContext =
-              new TraceContext(
-                  this, userAtomicReference.get(), hub.getOptions(), this.getSamplingDecision());
-        }
-        return this.traceContext;
-      }
+      updateBaggageValues();
+      return baggage.toTraceContext();
     } else {
       return null;
     }
   }
 
+  private void updateBaggageValues() {
+    synchronized (this) {
+      if (baggage.isMutable()) {
+        final AtomicReference<User> userAtomicReference = new AtomicReference<>();
+        hub.configureScope(
+            scope -> {
+              userAtomicReference.set(scope.getUser());
+            });
+        baggage.setValuesFromTransaction(
+            this, userAtomicReference.get(), hub.getOptions(), this.getSamplingDecision());
+        baggage.freeze();
+      }
+    }
+  }
+
   @Override
-  public @Nullable BaggageHeader toBaggageHeader() {
-    final TraceContext traceContext = traceContext();
-    if (hub.getOptions().isTraceSampling() && traceContext != null) {
-      final Baggage baggage = traceContext.toBaggage(hub.getOptions().getLogger());
-      return new BaggageHeader(baggage);
+  public @Nullable BaggageHeader toBaggageHeader(@Nullable List<String> thirdPartyBaggageHeaders) {
+    if (hub.getOptions().isTraceSampling()) {
+      updateBaggageValues();
+
+      return BaggageHeader.fromBaggageAndOutgoingHeader(baggage, thirdPartyBaggageHeaders);
     } else {
       return null;
     }
@@ -496,6 +513,27 @@ public final class SentryTracer implements ITransaction {
     return this.root.getData(key);
   }
 
+  @Override
+  public void setMeasurement(final @NotNull String name, final @NotNull Number value) {
+    if (root.isFinished()) {
+      return;
+    }
+
+    this.measurements.put(name, new MeasurementValue(value, null));
+  }
+
+  @Override
+  public void setMeasurement(
+      final @NotNull String name,
+      final @NotNull Number value,
+      final @NotNull MeasurementUnit unit) {
+    if (root.isFinished()) {
+      return;
+    }
+
+    this.measurements.put(name, new MeasurementValue(value, unit.apiName()));
+  }
+
   public @Nullable Map<String, Object> getData() {
     return this.root.getData();
   }
@@ -589,6 +627,12 @@ public final class SentryTracer implements ITransaction {
   @NotNull
   AtomicBoolean isFinishTimerRunning() {
     return isFinishTimerRunning;
+  }
+
+  @TestOnly
+  @NotNull
+  Map<String, MeasurementValue> getMeasurements() {
+    return measurements;
   }
 
   private static final class FinishStatus {