+++ sentry/src/test/java/io/sentry/MainEventProcessorTest.kt
@@ -32,8 +32,9 @@ class MainEventProcessorTest {
             dist = "dist"
             sdkVersion = SdkVersion("test", "1.2.3")
         }
+        val hub = mock<IHub>()
         val getLocalhost = mock<InetAddress>()
-        val sentryTracer = SentryTracer(TransactionContext("", ""), mock())
+        lateinit var sentryTracer: SentryTracer
         private val hostnameCacheMock = Mockito.mockStatic(HostnameCache::class.java)
 
         fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap(), sendDefaultPii: Boolean? = null, serverName: String? = "server", host: String? = null, resolveHostDelay: Long? = null, hostnameCacheDuration: Long = 10, proguardUuid: String? = null): MainEventProcessor {
@@ -55,6 +56,9 @@ class MainEventProcessorTest {
                 }
                 host
             }
+            whenever(hub.options).thenReturn(sentryOptions)
+            sentryTracer = SentryTracer(TransactionContext("", ""), hub)
+
             val hostnameCache = HostnameCache(hostnameCacheDuration) { getLocalhost }
             hostnameCacheMock.`when`<Any> { HostnameCache.getInstance() }.thenReturn(hostnameCache)